Refactored show-time-gap functionality.
Instead of using a popup editor, edit time gap value directly on the top-pane for each file.
